everything renders
