+++ src/components/_addEmployeeModal/View.js
@@ -1,67 +1,26 @@
-import Modal from "../modal/Modal";
 import {useState} from "react";
-import FormFields from "./FormFields";
-import Departments from "./utils/Departments";
+import AddEmployeeModal from "./AddEmployeeModal";
 
-const AddEmployeeModal = ({handleClose, isOpen}) => {
-    const [email, setEmail] = useState("")
-    const [password, setPassword] = useState("")
-    const [name, setName] = useState("")
-    const [surname, setSurname] = useState("")
-    const [department, setDepartment] = useState(Departments[0])
-    const [salary, setSalary] = useState(0)
-    const [stack, setStack] = useState([])
+const OpenModalButton = ({isLoading, actions}) => {
+    const [isOpen, setIsOpen] = useState(false);
 
-    const resetState = () => {
-        setEmail("")
-        setPassword("")
-        setName("")
-        setSurname("")
-        setDepartment(Departments[0])
-        setSalary(0)
-        setStack([])
+    const openModal = () => {
+        setIsOpen(true)
     }
 
-
-    const formFields = FormFields({
-        email: {
-            value: email,
-            setter: setEmail
-        },
-        password: {
-            value: password,
-            setter: setPassword
-        },
-        name: {
-            value: name,
-            setter: setName
-        },
-        surname: {
-            value: surname,
-            setter: setSurname
-        },
-        department: {
-            value: department,
-            setter: setDepartment
-        },
-        salary: {
-            value: salary,
-            setter: setSalary
-        },
-        stack: {
-            value: stack,
-            setter: setStack
-        }
-    })
+    const closeModal = () => {
+        setIsOpen(false)
+    }
 
     return (
-        <>
-            <Modal handleClose={() => {
-                resetState()
-                handleClose()
-            }} isOpen={isOpen} items={formFields} />
-        </>
+      <>
+          <button onClick={openModal}>
+              Click to Open Modal
+          </button>
+
+          <AddEmployeeModal handleClose={closeModal} isOpen={isOpen} isLoading={isLoading} actions={actions} />
+      </>
     )
 }
 
-export default AddEmployeeModal
+export default OpenModalButton
