+++ src/components/_addEmployeeModal/index.js
@@ -1,0 +1,4 @@
+import Container from "./Container";
+import View from "./View";
+
+export default Container(View)
